Fix dockerignore, shared package, proxy dedup, CI test gap, and document token storage
- Add .dockerignore to exclude node_modules, .git, .env, backups, etc from build context (#164)
- Wire packages/shared into the pnpm workspace and use it for real (#160)
- Extract duplicated proxy/trust-proxy parsing from apps/api and apps/addon into @cataloggy/shared (#161)
- Add a Test step to dockerpublish.yml so failing tests block image publish (#166)
- Document the localStorage token storage tradeoff as a known security limitation in the README (#180)
