+++ .dockerignore
@@ -1,0 +1,10 @@
+node_modules
+.git
+.env
+.env.*
+!.env.example
+backups/
+.claude/
+*.log
+coverage
+*.tsbuildinfo
