+++ README.md
@@ -175,6 +175,13 @@ To schedule automatic backups, add a cron entry that runs `backup.sh` on a sched
 
 In addition to full database backups, Settings → Data lets you export your lists, watch history, series progress, and ratings as a single JSON file, and re-import it later (e.g. after a fresh install, or to migrate to a new instance). The API also accepts CSV watch-history imports (columns: `imdbId`, `type`, `watchedAt`, with optional `season`/`episode`) for importing history exported from other tools.
 
+## Security
+
+Cataloggy is designed for self-hosting on a trusted local network (LAN), not for direct exposure to the internet. Known limitations:
+
+- The web client stores its API bearer token in `localStorage`. Any JavaScript running on the page (e.g. via an XSS vulnerability in a dependency) could read this token. Since the token only grants access to your own local API, this is an acceptable tradeoff for a LAN-only app, but it is not suitable for an internet-facing deployment without further hardening (e.g. a Content-Security-Policy header, or migrating to httpOnly session cookies backed by a session store).
+- If you do expose Cataloggy beyond your LAN, put it behind a reverse proxy with TLS and consider implementing a proper session-based auth flow.
+
 ## Useful Commands
 
 - Run apps in dev mode from root:
