+++ apps/addon/Dockerfile
@@ -8,11 +8,14 @@ COPY package.json pnpm-lock.yaml pnpm-workspace.yaml tsconfig.base.json eslint.c
 COPY apps/addon/package.json ./apps/addon/package.json
 COPY apps/api/package.json ./apps/api/package.json
 COPY apps/web/package.json ./apps/web/package.json
+COPY packages/shared/package.json ./packages/shared/package.json
 
 RUN pnpm install --frozen-lockfile
 
 COPY apps/addon ./apps/addon
+COPY packages/shared ./packages/shared
 
+RUN pnpm --filter @cataloggy/shared build
 RUN cd apps/addon && pnpm exec tsc -p tsconfig.json
 RUN addgroup -S app && adduser -S app -G app && chown -R app:app /app
 USER app
