+++ apps/addon/package.json
@@ -11,6 +11,7 @@
     "lint": "eslint src"
   },
   "dependencies": {
+    "@cataloggy/shared": "workspace:*",
     "@fastify/rate-limit": "^11.0.0",
     "fastify": "^5.1.0"
   },
