+++ apps/api/Dockerfile
@@ -8,12 +8,15 @@ COPY package.json pnpm-lock.yaml pnpm-workspace.yaml tsconfig.base.json eslint.c
 COPY apps/api/package.json ./apps/api/package.json
 COPY apps/addon/package.json ./apps/addon/package.json
 COPY apps/web/package.json ./apps/web/package.json
+COPY packages/shared/package.json ./packages/shared/package.json
 
 RUN pnpm install --frozen-lockfile
 
 COPY apps/api ./apps/api
+COPY packages/shared ./packages/shared
 
 RUN pnpm exec prisma generate --schema=apps/api/prisma/schema.prisma
+RUN pnpm --filter @cataloggy/shared build
 RUN pnpm --filter @cataloggy/api build
 RUN addgroup -S app && adduser -S app -G app && chown -R app:app /app
 USER app
