+++ apps/api/package.json
@@ -14,6 +14,7 @@
     "prisma:push": "prisma db push"
   },
   "dependencies": {
+    "@cataloggy/shared": "workspace:*",
     "@fastify/rate-limit": "^11.0.0",
     "@prisma/client": "^5.22.0",
     "fastify": "^5.1.0",
