+++ apps/api/src/index.ts
@@ -1,5 +1,6 @@
 import Fastify from "fastify";
 import rateLimit from "@fastify/rate-limit";
+import { parseProxyPathPrefixes, normalizeProxyPath, parseTrustProxy } from "@cataloggy/shared";
 import { prisma } from "./lib/prisma.js";
 import { applyCorsHeaders } from "./lib/cors.js";
 import { verifyToken } from "./lib/auth.js";
@@ -36,47 +37,12 @@ const TRAKT_POLL_INTERVAL_SEC = Number(process.env.TRAKT_POLL_INTERVAL_SEC ?? 30
 const AI_REFRESH_INTERVAL_SEC = Number(process.env.AI_REFRESH_INTERVAL_SEC ?? 86400);
 const NOTIFICATION_CHECK_INTERVAL_SEC = Number(process.env.NOTIFICATION_CHECK_INTERVAL_SEC ?? 3600);
 
-const parseProxyPathPrefixes = (raw: string | undefined, fallback: readonly string[]) => {
-  const parsed = (raw ?? "")
-    .split(",")
-    .map((prefix) => prefix.trim())
-    .filter(Boolean)
-    .map((prefix) => (prefix.startsWith("/") ? prefix : `/${prefix}`));
-
-  return parsed.length > 0 ? parsed : [...fallback];
-};
-
 const PROXY_PATH_PREFIXES = parseProxyPathPrefixes(process.env.PROXY_PATH_PREFIXES, ["/api"] as const);
-
-const stripProxyPrefix = (url: string, prefix: string) => {
-  if (url === prefix) return "/";
-  if (!url.startsWith(`${prefix}/`)) return null;
-  return url.slice(prefix.length) || "/";
-};
-
-const normalizeProxyPath = (rawUrl: string) => {
-  for (const prefix of PROXY_PATH_PREFIXES) {
-    const stripped = stripProxyPrefix(rawUrl, prefix);
-    if (stripped) return stripped;
-  }
-  return rawUrl;
-};
-
-// Only trust X-Forwarded-* headers from explicitly configured proxies, so
-// request.ip (used as the rate-limit key) can't be spoofed by clients when
-// there's no reverse proxy in front of this service.
-const parseTrustProxy = (raw: string | undefined): boolean | string[] | undefined => {
-  if (!raw) return undefined;
-  const trimmed = raw.trim();
-  if (trimmed === "true") return true;
-  if (trimmed === "false") return false;
-  return trimmed.split(",").map((entry) => entry.trim()).filter(Boolean);
-};
 
 const app = Fastify({
   logger: true,
   trustProxy: parseTrustProxy(process.env.TRUST_PROXY),
-  rewriteUrl: (request) => normalizeProxyPath(request.url ?? "/"),
+  rewriteUrl: (request) => normalizeProxyPath(request.url ?? "/", PROXY_PATH_PREFIXES),
 });
 
 // ─── Rate limiting ───
